Moves helper classes and functions to new module
Moves *Manager classes and pack/unpack functions to the persistence_helpers module
